+++ packages/adl/compiler/options.ts
@@ -5,4 +5,11 @@ export interface CompilerOptions {
   nostdlib?: boolean;
   noEmit?: boolean;
   serviceCodePath?: string;
+  /**
+   * When true, indicates that a compilation is being performed for live
+   * analysis in the language server.
+   *
+   * This currently disables execution of decorators and onBuild handlers.
+   */
+  designTimeBuild?: boolean;
 }